+++ modules/buildkite_elastic_stack/stack.tf
@@ -44,6 +44,7 @@ resource "aws_cloudformation_stack" "stack" {
     BuildkiteQueue                 = var.buildkite_queue
     ECRAccessPolicy                = "poweruser"
     EnableDockerUserNamespaceRemap = "false"
+    EnableExperimentalLambdaBasedAutoscaling = var.enable_experimental_lambda_based_autoscaling
     InstanceType                   = var.instance_type
     KeyName                        = var.key_name
     ManagedPolicyARN               = var.managed_policy_arn
@@ -61,7 +62,7 @@ resource "aws_cloudformation_stack" "stack" {
   }
 
   lifecycle {
-    ignore_changes = [parameters.BuildkiteAgentToken]
+    ignore_changes = [parameters["BuildkiteAgentToken"]]
   }
 }